Fix bug where single email is split into chars
I join the list of email addresses by a comma, turning them all into a
single string. However I don't check to see if the email addresses
supplied are a list of strings or a single string. When it's a single
string it's still run through the `','.join()` function, which is not
what we want. It should only do that for lists.